Add files via upload
Dr. Newby, I attempted to do the graph we discussed, but my lists kept coming up saying that there was an unexpected character in what I was trying to plot. The number of where it was would change and I could not fix it. Instead, I had to make this graph, which is significantly less what I had in mind. I thought I should explain why the “trend” was very plain.